requirements.txt
matplotlib==3.5.2
pandas==1.4.3
Pillow==9.2.0
requests==2.28.1
streamlit==1.11.0
streamlit_lottie==0.0.3
